requirements.txt
aniso8601==8.0.0
astroid==2.3.3
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
coverage==4.5.4
coveralls==1.9.2
docopt==0.6.2
Flask==1.1.1
Flask-RESTful==0.3.7
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.3.0
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psycopg2-binary==2.8.4
py==1.8.0
pylint==2.4.4
pyparsing==2.4.5
pytest==5.3.1
pytest-cov==2.8.1
pytest-pylint==0.14.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
wrapt==1.11.2
